debug

+++ MuonAnalysis/python/DATA.py
@@ -46,10 +46,11 @@ def prepareSet( data, label_dict):
 	
 	x_test = data.to_numpy()
 	y_test = target.to_numpy()
-	y_test = y_test.astype(int)
+
 	
 
-	y_test = np.array([np.array(i).astype(int) for i in y_test])
+	y_test = np.array([np.array(i) for i in y_test])
+	y_test = y_test.astype(int)
 	pt_test = pt_test.to_numpy()
 	return x_test,y_test,pt_test
